refactored isostericity to its own package. all tests pass.
