+++ tests/test_isosteric/__init__.py
@@ -1,0 +1,2 @@
+from test_isostericity import *
+from test_isostericity_matrix import *
